requirements.txt
aiohttp
feedparser
pyyaml
